More extraction.
Get ready to test.
