+++ lib/generate.js
@@ -1,13 +1,7 @@
 var Metalsmith = require('metalsmith')
-var inquirer = require('inquirer')
 var async = require('async')
 var render = require('consolidate').handlebars.render
 var path = require('path')
-
-var promptInquirerTypeMapping = {
-  string: 'input',
-  boolean: 'confirm'
-}
 
 /**
  * Template in place plugin.
@@ -37,49 +31,7 @@ function renderTemplateFiles (files, metalsmith, done) {
   }
 }
 
-module.exports = function (name) {
-  var options = require('../lib/options')(name)
-
-  /**
-   * Prompt plugin.
-   *
-   * @param {Object} files
-   * @param {Metalsmith} metalsmith
-   * @param {Function} done
-   */
-
-  function ask (files, metalsmith, done) {
-    var opts = options(metalsmith._directory + '/..')
-
-    var prompts = Object.keys(opts.schema)
-    var metalsmithMetadata = metalsmith.metadata()
-
-    async.eachSeries(prompts, run, done)
-
-    function run (key, done) {
-      var prompt = opts.schema[key]
-
-      inquirer.prompt([{
-        type: promptInquirerTypeMapping[prompt.type] || prompt.type,
-        name: key,
-        message: prompt.label || key,
-        default: prompt.default,
-        choices: prompt.choices || []
-      }], function (answers) {
-        if (Array.isArray(answers[key])) {
-          metalsmithMetadata[key] = {}
-          answers[key].forEach(function (multiChoiceAnswer) {
-            metalsmithMetadata[key][multiChoiceAnswer] = true
-          })
-        } else {
-          metalsmithMetadata[key] = answers[key]
-        }
-
-        done()
-      })
-    }
-  }
-
+module.exports = function (ask) {
   /**
    * Generate a template given a `src` and `dest`.
    *
